recommendation-engine/requirements.txt
pyspark
psycopg2-binary
python-dotenv
fastapi
uvicorn